lib/settings: rename slidesSettings.js to presentationSettings.js
