+++ lib/settings/index.js
@@ -3,5 +3,5 @@
 */
 
 module.exports = {
-  slidesSettings          : require('./slidesSettings')
+  presentationSettings          : require('./presentationSettings')
 }
